fix: переносит форму логина в компонент
